update | Added Multi User views | refactoring
